+++ castor/db/ora/OraStageOutRequestCnv.cpp
@@ -17,7 +17,7 @@
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA.
  *
- * @(#)$RCSfile: OraStageOutRequestCnv.cpp,v $ $Revision: 1.2 $ $Release$ $Date: 2004/10/05 13:37:28 $ $Author: sponcec3 $
+ * @(#)$RCSfile: OraStageOutRequestCnv.cpp,v $ $Revision: 1.3 $ $Release$ $Date: 2004/10/07 09:46:49 $ $Author: sponcec3 $
  *
  * 
  *
@@ -339,7 +339,7 @@ void castor::db::ora::OraStageOutRequestCnv::updateRep(castor::IAddress* address
       }
       // Dealing with client
       {
-        u_signed64 clientId = rset->getInt(11);
+        u_signed64 clientId = (unsigned long long)rset->getDouble(11);
         castor::db::DbAddress ad(clientId, " ", 0);
         if (0 != clientId &&
             0 != obj->client() &&
@@ -534,7 +534,7 @@ castor::IObject* castor::db::ora::OraStageOutRequestCnv::createObj(castor::IAddr
     // create the new Object
     castor::stager::StageOutRequest* object = new castor::stager::StageOutRequest();
     // Now retrieve and set members
-    object->setFlags(rset->getDouble(1));
+    object->setFlags((unsigned long long)rset->getDouble(1));
     object->setUserName(rset->getString(2));
     object->setEuid(rset->getInt(3));
     object->setEgid(rset->getInt(4));
@@ -543,9 +543,9 @@ castor::IObject* castor::db::ora::OraStageOutRequestCnv::createObj(castor::IAddr
     object->setMachine(rset->getString(7));
     object->setProjectName(rset->getString(8));
     object->setOpenmode(rset->getInt(9));
-    object->setId(rset->getDouble(10));
+    object->setId((unsigned long long)rset->getDouble(10));
     newlyCreated[object->id()] = object;
-    u_signed64 clientId = rset->getInt(11);
+    u_signed64 clientId = (unsigned long long)rset->getDouble(11);
     IObject* objClient = cnvSvc()->getObjFromId(clientId, newlyCreated);
     object->setClient(dynamic_cast<castor::IClient*>(objClient));
     object->setStatus((enum castor::stager::RequestStatusCodes)rset->getInt(12));
@@ -612,7 +612,7 @@ void castor::db::ora::OraStageOutRequestCnv::updateObj(castor::IObject* obj,
     // Now retrieve and set members
     castor::stager::StageOutRequest* object = 
       dynamic_cast<castor::stager::StageOutRequest*>(obj);
-    object->setFlags(rset->getDouble(1));
+    object->setFlags((unsigned long long)rset->getDouble(1));
     object->setUserName(rset->getString(2));
     object->setEuid(rset->getInt(3));
     object->setEgid(rset->getInt(4));
@@ -621,10 +621,10 @@ void castor::db::ora::OraStageOutRequestCnv::updateObj(castor::IObject* obj,
     object->setMachine(rset->getString(7));
     object->setProjectName(rset->getString(8));
     object->setOpenmode(rset->getInt(9));
-    object->setId(rset->getDouble(10));
+    object->setId((unsigned long long)rset->getDouble(10));
     alreadyDone[obj->id()] = obj;
     // Dealing with client
-    u_signed64 clientId = rset->getInt(11);
+    u_signed64 clientId = (unsigned long long)rset->getDouble(11);
     if (0 != object->client() &&
         (0 == clientId ||
          object->client()->id() != clientId)) {
